requirements/requirements.txt
six
pyyaml
pexpect
